fix: CallRecord.create
Change the expose to CallRecord.create.
And fix tersifing it.
