+++ src/CallEntry.ts
@@ -1,0 +1,20 @@
+import { CallRecord } from './CallRecord'
+
+export interface CallEntry extends Promise<any> {
+  inputs: any[],
+  /**
+   * Synchronous result.
+   */
+  output: any,
+  /**
+   * Synchronous error got thrown.
+   */
+  error: any,
+  getCallRecord(): Promise<CallRecord>
+}
+
+// export const CallEntry = {
+//   create(callRecord: CallRecord) {
+//     return {}
+//   }
+// }
